Adjustments

+++ src/tasks/tasks.service.ts
@@ -16,12 +16,14 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ){}
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-        return await this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+        return await this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id }});
 
         if(!found) {            
             throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -30,8 +32,10 @@ export class TasksService {
         return found;
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);   
+    async deleteTask(
+        id: number,
+        user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id});   
 
         if(result.affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -39,8 +43,12 @@ export class TasksService {
         
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task>{
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number, 
+        status: TaskStatus,
+        user: User): Promise<Task>{
+
+        const task = await this.getTaskById(id, user);
         task.status = status;
 
         await task.save();
